JWT+Auth0

+++ src/main/java/com/nestenote/notenest/service/NoteService.java
@@ -4,6 +4,7 @@ import com.nestenote.notenest.model.Note;
 import com.nestenote.notenest.repository.NoteRepository;
 import com.nestenote.notenest.dto.NoteRequestDTO;
 import com.nestenote.notenest.dto.NoteResponseDTO;
+import com.nestenote.notenest.exception.NoteNotFoundException;
 import com.nestenote.notenest.exception.ValidationException;
 import org.springframework.stereotype.Service;
 
@@ -23,10 +24,11 @@ public class NoteService {
     }
 
     // converts user inputs restricted by DTOs to a full database entity
-    private Note mapToEntity(NoteRequestDTO dto) {
+    private Note mapToEntity(NoteRequestDTO dto,String userID) {
         Note note = new Note();
         note.setTitle(dto.getTitle());
         note.setContent(dto.getContent());
+        note.setUserId(userID);
         note.setCreatedAt(LocalDateTime.now());
         return note;
     }
@@ -42,23 +44,29 @@ public class NoteService {
     }
 
     // Get all notes
-    public List<NoteResponseDTO> getAllNotes() {
-        return noteRepository.findAll()
+    public List<NoteResponseDTO> getAllUserNotes(String userID) {
+        return noteRepository.findByUserID(userID)
                 .stream()
                 .map(this::mapToDTO)
                 .collect(Collectors.toList());
     }
 
     // Get one note by ID
-    public Optional<NoteResponseDTO> getNoteById(Long id) {
-        return noteRepository.findById(id)
-                .map(this::mapToDTO);
+    public Optional<NoteResponseDTO> getNoteByID(Long id, String userID) {
+        Note note = noteRepository.findById(id)
+                 .filter(n -> n.getUserId().equals(userID))
+                .orElseThrow(() -> new NoteNotFoundException(id));
+                
+        return Optional.of(mapToDTO(note));
     }
 
     // Create a new note
-    public NoteResponseDTO createNote(NoteRequestDTO dto) {
+    public NoteResponseDTO createNote(NoteRequestDTO dto,String userID) {
+        if(dto.getTitle() == null || dto.getTitle().isEmpty()) {
+            throw new ValidationException("Title cannot be empty");
+        }
         // creates a full note object from dto
-        Note note = mapToEntity(dto);
+        Note note = mapToEntity(dto, userID);
         // saves full note into entity
         Note saved = noteRepository.save(note);
         // returns the safe DTO note version
@@ -66,8 +74,10 @@ public class NoteService {
     }
 
     // Update an existing note
-    public Optional<NoteResponseDTO> updateNote(Long id, NoteRequestDTO dto) {
-        return noteRepository.findById(id).map(note -> {
+    public Optional<NoteResponseDTO> updateNote(Long id, NoteRequestDTO dto, String userID) {
+        return noteRepository.findById(id)
+        .filter(note -> note.getUserId().equals(userID))
+        .map(note -> {
             note.setTitle(dto.getTitle());
             note.setContent(dto.getContent());
             Note updated = noteRepository.save(note);
@@ -76,8 +86,11 @@ public class NoteService {
     }
 
     // Delete a note by ID
-    public boolean deleteNote(Long id) {
+    public boolean deleteNote(Long id,String userID) {
         return noteRepository.findById(id).map(note -> {
+            if(!note.getUserId().equals(userID)) {
+                throw new NoteNotFoundException(id);
+            }
             noteRepository.delete(note);
             return true;
         }).orElse(false);
